requirements.txt
requests==2.31.0
azure-functions==1.18.0
azure-storage-blob==12.19.0
